#50 - refactor code
