+++ server/src/auth/strategies/jwt.strategy.ts
